+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+
+0.8.0 (2025-02-01)
+------------------
 * Added Added methods update_checklists() and update_checklist() to the APILoader
   for updating existing checklists.
 * Added Checklist.started to record the date and time the checklist was started.